Add single coefficient plane support. (RAM optimization)

+++ src/CachedTiler.cpp
@@ -42,8 +42,8 @@ CachedTiler::CachedTiler (size_t display_width, size_t display_height,
 #else
     display_ = new GlNddiDisplay(fvDimensions,                  // framevolume dimensional sizes
                                  display_width, display_height, // display size
-                                 1,                             // number of coefficient planes in display
-                                 3,                             // input vector size (x, y, and z)
+                                 (unsigned int)1,               // number of coefficient planes in display
+                                 (unsigned int)3,               // input vector size (x, y, and z)
                                  globalConfiguration.headless);
 #endif
