+++ src/MultiDctTiler.cpp
@@ -92,10 +92,10 @@ MultiDctTiler::MultiDctTiler(size_t display_width, size_t display_height, size_t
                                  3,                             // Input vector size (x, y, 1)
                                  globalConfiguration.headless);
 #else
-    display_ = new GlNddiDisplay(fvDimensions,                  // framevolume dimensional sizes
-                                 display_width, display_height, // display size
-                                 FRAMEVOLUME_DEPTH,             // Number of coefficient planes
-                                 3,                             // Input vector size (x, y, 1)
+    display_ = new GlNddiDisplay(fvDimensions,                    // framevolume dimensional sizes
+                                 display_width, display_height,   // display size
+                                 (unsigned int)FRAMEVOLUME_DEPTH, // Number of coefficient planes
+                                 (unsigned int)3,                 // Input vector size (x, y, 1)
                                  globalConfiguration.headless);
 #endif
